+++ ghost/core/core/server/api/endpoints/oembed.js
@@ -1,25 +1,4 @@
-const config = require('../../../shared/config');
-const externalRequest = require('../../lib/request-external');
-
-const OEmbed = require('@tryghost/oembed-service');
-const oembed = new OEmbed({config, externalRequest});
-
-const NFT = require('../../services/nft-oembed');
-const nft = new NFT({
-    config: {
-        apiKey: config.get('opensea').privateReadOnlyApiKey
-    }
-});
-
-const Twitter = require('../../services/twitter-embed');
-const twitter = new Twitter({
-    config: {
-        bearerToken: config.get('twitter').privateReadOnlyToken
-    }
-});
-
-oembed.registerProvider(nft);
-oembed.registerProvider(twitter);
+const oembed = require('../../services/oembed');
 
 module.exports = {
     docName: 'oembed',
