+++ ghost/core/core/server/services/oembed/index.js
@@ -1,0 +1,1 @@
+module.exports = require('./service');
